requirements.txt
flask==3.0.3
python-magic==0.4.27
